Added Loading of Questions

+++ src/api/allTests/CutOff.kt
@@ -2,6 +2,6 @@ package api.allTests
 
 data class CutOff(
     var category: String?,
-    var lowerBound: Int?,
-    var upperBound: Int?
+    var lowerBound: Double?,
+    var upperBound: Double?
 )
