+++ src/api/question/CutOff.kt
@@ -2,6 +2,6 @@ package api.question
 
 data class CutOff(
     var category: String?,
-    var lowerBound: Int?,
-    var upperBound: Int?
+    var lowerBound: Double?,
+    var upperBound: Double?
 )
